# edge/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Police Incident Management System API
  description: |
    REST API for the Police Incident Management System. This is an event-driven system where 
    all operations produce events to Kafka. Events represent requests/commands from the edge, 
    not state changes.
    
    ## Architecture
    - Edge servers receive HTTP requests and produce events to Kafka
    - Events represent requests/commands, not state changes
    - No state reconstruction in edge layer
    - All operations are asynchronous via Kafka events
    
  version: 1.0.0
  contact:
    name: Police System API Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.policesystem.example.com/api/v1
    description: Production server
...
paths:
...
  /officers:
    post:
      tags:
        - Officers
      summary: Register a new officer
      description: Registers a new police officer and produces a RegisterOfficerRequested event to Kafka
      operationId: registerOfficer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterOfficerRequest'
      responses:
        '201':
          description: Officer registration request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Officer with badge number already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /officers/{badgeNumber}:
    put:
      tags:
        - Officers
      summary: Update officer information
      description: Updates officer information and produces an UpdateOfficerRequested event to Kafka
      operationId: updateOfficer
      parameters:
        - $ref: '#/components/parameters/BadgeNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfficerRequest'
      responses:
        '200':
          description: Officer update request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficerUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Officers
      summary: Change officer status
      description: Changes officer status and produces a ChangeOfficerStatusRequested event to Kafka
      operationId: changeOfficerStatus
      parameters:
        - $ref: '#/components/parameters/BadgeNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      responses:
        '200':
          description: Officer status change request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficerStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /officers/batch:
    post:
      tags:
        - Officers
      summary: Batch register officers
      description: Registers multiple officers in one call with partial failure handling
      operationId: batchRegisterOfficers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreateOfficersRequest'
      responses:
        '201':
          description: Batch officer registration completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================
  # Vehicles
  # ============================================
...
  /query/{domain}/{id}/full:
    get:
      tags:
        - Query
      summary: Get full resource from projections
      description: Retrieves a complete resource with related data from projection services
      operationId: getFullResource
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
            example: "incident"
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "INC-001"
        - name: fields
          in: query
          required: false
          schema:
            type: string
            example: "id,description,status"
          description: "Comma-separated list of fields to include"
      responses:
        '200':
          description: Resource retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
        '404':
          $ref: '#/components/responses/NotFound'
  /query/{domain}/{id}/exists:
    get:
      tags:
        - Query
      summary: Check resource existence
      description: Checks if a resource exists in projections
      operationId: checkResourceExists
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Existence check completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          exists:
                            type: boolean
  /query/{domain}:
    get:
      tags:
        - Query
      summary: List resources with filtering and pagination
      description: Lists resources from projections with optional filtering, pagination, and sorting
      operationId: listResources
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            example: "id,asc"
      responses:
        '200':
          description: Resources retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                          totalElements:
                            type: integer
                          totalPages:
                            type: integer
                          page:
                            type: integer
                          size:
                            type: integer

  # ============================================
  # Activities
  # ============================================
...
  /shifts/{shiftId}/officers/{badgeNumber}:
    put:
      tags:
        - OfficerShifts
      summary: Update officer shift
      description: Updates officer shift information and produces an UpdateOfficerShiftRequested event to Kafka
      operationId: updateOfficerShift
      parameters:
        - $ref: '#/components/parameters/ShiftId'
        - $ref: '#/components/parameters/BadgeNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfficerShiftRequest'
      responses:
        '200':
          description: Officer shift update request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    BadgeNumber:
      name: badgeNumber
      in: path
      required: true
      description: Officer badge number
      schema:
        type: string
        example: "12345"
...
    ShiftId:
      name: shiftId
      in: path
      required: true
      description: Shift identifier
      schema:
        type: string
        example: "SHIFT-001"
...
  schemas:
...
    OfficerStatus:
      type: string
      enum:
        - Active
        - On-Duty
        - Off-Duty
        - Suspended
        - Retired
      example: "Active"
...
    ShiftRoleType:
      type: string
      enum:
        - Regular
        - Supervisor
        - Trainee
        - Other
      example: "Regular"
...
    RegisterOfficerRequest:
      type: object
      required:
        - badgeNumber
        - firstName
        - lastName
        - rank
        - email
        - status
      properties:
        badgeNumber:
          type: string
          description: Unique badge number
          example: "12345"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        rank:
          type: string
          example: "Officer"
        email:
          type: string
          format: email
          example: "john.doe@police.gov"
        phoneNumber:
          type: string
          example: "555-0100"
        hireDate:
          type: string
          format: date
          example: "2020-01-15"
        status:
          $ref: '#/components/schemas/OfficerStatus'
    UpdateOfficerRequest:
      type: object
      properties:
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        rank:
          type: string
          example: "Sergeant"
        email:
          type: string
          format: email
          example: "john.doe@police.gov"
        phoneNumber:
          type: string
          example: "555-0100"
        hireDate:
          type: string
          format: date
          example: "2020-01-15"
...
    UpdateOfficerShiftRequest:
      type: object
      properties:
        shiftRoleType:
          $ref: '#/components/schemas/ShiftRoleType'
...
    ChangeStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "Active"

    # Response Schemas
    OfficerResponse:
      type: object
      properties:
        officerId:
          type: string
          example: "12345"
        badgeNumber:
          type: string
          example: "12345"
    OfficerUpdateResponse:
      type: object
      properties:
        badgeNumber:
          type: string
          example: "12345"
        message:
          type: string
          example: "Officer update request processed"
    OfficerStatusResponse:
      type: object
      properties:
        badgeNumber:
          type: string
          example: "12345"
        status:
          $ref: '#/components/schemas/OfficerStatus'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Request processed successfully"
        data:
          description: Response data (type varies by endpoint)
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
          nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: string
          example: ["badgeNumber is required"]
        validValues:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field names to their valid enum values (only present for enum validation errors)
          example:
            unitType: ["Single", "Team", "Squad"]
    ResponseMetadata:
      type: object
      properties:
        correlationId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: "Correlation ID for request tracing"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:40:00.000+00:00"
        version:
          type: string
          example: "1.0"
        links:
          type: object
          additionalProperties:
            type: string
          description: "HATEOAS-style links to related resources"
          example:
            self: "/api/v1/incidents/INC-001"
            dispatch: "/api/v1/incidents/INC-001/dispatch"
...
    BatchCreateOfficersRequest:
      type: object
      required:
        - officers
      properties:
        officers:
          type: array
          items:
            $ref: '#/components/schemas/RegisterOfficerRequest'
          minItems: 1
    BatchCreateResponse:
      type: object
      properties:
        created:
          type: array
          items:
            type: string
          example: ["BADGE-001", "BADGE-002"]
        failed:
          type: array
          items:
            $ref: '#/components/schemas/BatchFailure'
    BatchFailure:
      type: object
      properties:
        id:
          type: string
          example: "BADGE-003"
        reason:
          type: string
          example: "Officer with badge number already exists: BADGE-003"
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'